requirements.txt
Faker==21.0.0
Flask==3.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
stripe==7.10.0
